Women's Day Offer Challenge
The flow of challenge is as follows:
- The user obtains the coupon code 'WMNSDY2019' either through the
twitter account or by guessing the coupon code(knowing the fact that
the offer was on Women's day of 2019).
- When the user tries to use the coupon code, it is errored saying
'Invalid coupon.'
- The user will then modify their system time to be that of the
Womens Day(08 March 2019), after this using the coupon code will give
the user a discount of 75%.
- The user will then checkout(complete the purchase) to solve the
challenge.

Please let me know of suggestions and improvements. Thank You